Restructure of file/module layout to make more sense
