+++ src/rdf/parser.rs
@@ -197,56 +197,6 @@ impl ParsedTriples {
   }
 }
 
-pub fn parse_turtle(path: String) -> Vec<Triple> {
-  use rio_turtle::{TurtleParser, TurtleError};
-  use rio_api::parser::TriplesParser;
-  use rio_api::model::{
-    NamedOrBlankNode,
-    NamedNode,
-    Term,
-    Literal::{Simple, Typed, LanguageTaggedString}
-  };
-  use rio_api::model::Triple as RioTriple;
-  use std::io::BufReader;
-  use std::fs::File;
-
-  let mut ret_v: Vec<Triple> = Vec::new();
-
-  TurtleParser::new(BufReader::new(File::open(path.clone()).unwrap()), &format!("file:{}", path))
-    .unwrap()
-    .parse_all(&mut |t| {
-      match t {
-        RioTriple {
-          subject: NamedOrBlankNode::NamedNode(NamedNode { iri: s }),
-          predicate: NamedNode { iri: p },
-          object: o,
-        } => {
-          match o {
-            Term::Literal(Simple {
-              value: o
-            }) |
-            Term::Literal(Typed {
-              value: o,
-              datatype: _,
-            }) |
-            Term::Literal(LanguageTaggedString {
-              value: o,
-              language: _,
-            }) |
-            Term::NamedNode(NamedNode { iri: o }) => {
-              ret_v.push([s.to_string(), p.to_string(), o.to_string()]);
-            },
-            _ => {},
-          };
-        },
-        _ => {},
-      };
-      Ok(()) as Result<(), TurtleError>
-  });
-
-  ret_v
-}
-
 #[cfg(test)]
 mod unit_tests {
   use super::*;
