+++ src/rdf/query.rs
